reorg utils.ts to utils folder
