+++ projects/oneteme/jquery-highcharts/src/lib/directive/utils/chart-utils.ts
@@ -1,0 +1,33 @@
+import { CommonChart, Coordinate2D, XaxisType, YaxisType } from '@oneteme/jquery-core';
+
+/**
+ * Détermine le type de données pour l'axe X
+ */
+export function determineXAxisDataType(value: any): string {
+  if (value instanceof Date) {
+    return 'datetime';
+  } else if (typeof value === 'number') {
+    return 'numeric';
+  } else {
+    return 'category';
+  }
+}
+
+/**
+ * Détermine le type d'axe X en fonction des données
+ */
+export function getType<
+  X extends XaxisType,
+  Y extends YaxisType | Coordinate2D
+>(commonChart: CommonChart<X, Y>): string {
+  if (commonChart.series.length && commonChart.series[0].data.length) {
+    if (commonChart.continue) {
+      const x = (<CommonChart<X, Coordinate2D>>commonChart).series[0].data[0].x;
+      return determineXAxisDataType(x);
+    } else {
+      const categ = commonChart.categories[0];
+      return determineXAxisDataType(categ);
+    }
+  }
+  return 'datetime';
+}
